fixed bugs of email again

+++ IntertherapySiteBackend/src/main/java/com/Intertherapy/IntertherapySiteBackend/Treatment/TreatmentsRepository.java
@@ -2,5 +2,5 @@ package com.Intertherapy.IntertherapySiteBackend.Treatment;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TreatmentsRepository extends JpaRepository<TreatmentsClass,Long>{
+public interface TreatmentsRepository extends JpaRepository<TreatmentsClass,Integer>{
 }
